# backend/openapi.yaml
openapi: 3.0.0
info:
  title: BeachVillas User-Facing API
  version: 1.0.0
  description: >
    Definitive OpenAPI contract for all user-facing, data-related operations for
    the BeachVillas platform. Aligns strictly to the detailed PRD, FRD, DB
    schema, and UX datamap. JWT authentication, RESTful JSON responses, and
    comprehensive route coverage.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    AuthToken:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: integer
      required:
        - token
        - expires_at
    UserShort:
      type: object
      properties:
        user_id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        profile_photo_url:
          type: string
          nullable: true
        is_active:
          type: boolean
        is_verified_host:
          type: boolean
          nullable: true
        notification_settings:
          type: object
          additionalProperties: true
        payout_method_details:
          type: string
          nullable: true
      required:
        - user_id
        - name
        - email
        - role
        - is_active
        - notification_settings
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/UserShort"
        - type: object
          properties:
            about:
              type: string
              nullable: true
            phone:
              type: string
              nullable: true
            locale:
              type: string
              nullable: true
            created_at:
              type: integer
            updated_at:
              type: integer
          required:
            - created_at
            - updated_at
    UserUpdatePayload:
      type: object
      properties:
        name:
          type: string
        profile_photo_url:
          type: string
        about:
          type: string
        phone:
          type: string
        locale:
          type: string
        notification_settings:
          type: object
        payout_method_details:
          type: string
      required: []
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    PasswordResetSubmit:
      type: object
      properties:
        email:
          type: string
        token:
          type: string
        new_password:
          type: string
      required:
        - email
        - token
        - new_password
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
paths:
  /auth/signup:
    post:
      summary: Register new user (email/password or OAuth)
      operationId: auth_signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                provider:
                  type: string
                  nullable: true
                provider_token:
                  type: string
                  nullable: true
              required:
                - name
                - email
                - role
      responses:
        "200":
          description: Auth token with user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserShort"
                required:
                  - token
                  - user
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      summary: User login with email/password or OAuth provider
      operationId: auth_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                  nullable: true
                provider:
                  type: string
                  nullable: true
                provider_token:
                  type: string
                  nullable: true
              required:
                - email
      responses:
        "200":
          description: Auth token with user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserShort"
                required:
                  - token
                  - user
        "400":
          description: Bad login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      summary: Logout current user
      operationId: auth_logout
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logout successful, no content
  /auth/forgot-password:
    post:
      summary: Forgot password request (send reset link)
      operationId: auth_forgot_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "204":
          description: Reset email sent if user found; always 204 (no info leak)
  /auth/reset-password:
    post:
      summary: Submit new password with reset token
      operationId: auth_reset_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetSubmit"
      responses:
        "200":
          description: Password updated, return new token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "400":
          description: Bad reset request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me:
    get:
      summary: Get current user profile and settings
      operationId: me_get
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      summary: Update own profile info/settings
      operationId: me_update
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdatePayload"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...